Fixed RPC configuration
Due to the user_variables.yml file existing in both OSA
and RPC, we needed to adjust how we configured the RPC config
files. This is accomplished by using update-yaml script in RPC
to merge the 2 files. I also removed the cinder volume config and
swift volume lvs config out of the prepare tag as these should be run
seperate if needed