Add script to calculate upgrade paths
rpc-openstack supports two types of upgrade, major upgrades using
upgrade.sh and minor/patch upgrades where running a newer version of the
playbooks should result in the same/compatible configuration deployed.

This commit adds a script to determine the supported upgrade paths for
any commit. It assumes that x.0.z does not support major upgrades and
that minor version can only upgrade between contiguous releases.

master is calculated to be the next major release if the
closest openstack-ansible tag to the current SHA is on the next major
release to rpc-o. If it is not the next major it is assumed to be the
next minor release.