Report top level job instead of immediate parent
Currently there is a three level chain for aio gate jobs:
*-PR --> Docs-Test --> RPC-AIO.

The build summary page reports the Docs-Test job as the parent for AIO
jobs, this is not very helpful as it doesn't communicate whether the job
is an upgrade, or ceph/swift etc. Reporting the top level job ensures
this information is reported, however the hierachy is modified.