Show entire job hierachy in build summary.
There are situations in which the whole hierachy is useful, for example
in the PR jobs, the top level job is useful to show which type of job is
running, whereas in matrix builds this information is in the name of an
itermediate job.

This commit removes the job/parent job fields and replaces them with a
list of jobs starting from the root trigger.