RPC AIO: Fail Fast
This commit adds "set -e" to the aio build script, so that the script
will halt on the first failure that is encountered. Previously the
script continued through failure but stored a binary success/fail value
for each phase. As the script will now halt on failure it is no longer
necessary to store those results. Success lines have been added to the
log to give an indication of the last successfull step when diagnosing
failures.