Fix subshell interaction with -e
When -e is set, bash will:
"Exit immediately if a simple command exits with a non-zero status,
unless the subshell is part of the command list immediately following a
`while' or `until' keyword, part of the test in an `if' statement, part
of the right-hand-side of a `&&' or `||' list, or if the command's
return status is being inverted using `!'"

So using -e in a subshell is not reliable especially if the subshell is
part of larger logically combined statement.

This patch removes the reliance on -e in the subshell statements by
replacing ; with &&.

This patch also sets pipefail to catch errors that may have been masked
by pipes.