Add default value for possibly undefined var
if [[ ! -z "${ghprbTargetBranch}" ]]; then

This line attempts to test whether the var is set, however if the var is
not set, than trying to read it will cause a failure under -u. This
patch adds a default value to the variable so that it can be read but
the if statement will still evaluate corectly.

Contains / Depends on #613