Fix aio_build_script.sh with unset vars
An issue appears when ghprbAuthorRepoGitUrl or ghprbActualCommit
are not set, due to the presence of set -u in L3.

The condition used for horizon extensions was causing this, because
the variables above are only set for PRs, not periodic jobs.

This commit fixes the condition used, by nesting tests and
testing head of the PR for the presence of the horizon extensions
dir rather than testing the merged tree.

Connects rcbops/u-suk-dev#260

Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>
Co-Authored-By: Hugh Saunders <hugh.saunders@rackspace.co.uk>