Auto Select Mirror
If UBUNTU_MIRROR is unset or set to "auto"
Mirror list is from OSA, which uses the mirrors run by openstack infra.
Selection criteria is latency (time to load index page).

Also allows unauthenticated packages when UBUNTU_MIRROR is set to
"auto". The security role checks for the use of unauthenticated repos, so
those checks are disabled.

Connects rcbops/u-suk-dev#300