Catch exceptions when reading jobs
Jobs may have incomplete logs/configs if they are aborted at a bad time.
Build summary should not fail if one log is unreadable, it should skip
and continue.

This patch adds primative exception handling and logging to show which
jobs have files that cannot be read. Note that logging is to stderr
because the html is written to stdout.

Connects rcbops/u-suk-dev#311