Add JJB Job test for Jenkins RPC PRs
Adds a job that tests changes to JJB job definitions.
This is a simple test to ensure that proposed changes to compile.

Modifies the run_jjb script to add an operation paramter, this allows
the script to be used for checks as well as for syncing repo definitions
with jenkins.

Connects rcbops/u-suk-dev#345
Depends (includes) rcbops/jenkins-rpc#639