JJB Upgrade Matrix and Remove Jclouds Dep
The previous upgrade matrix job was a chain of Upgrade-Periodic ->
Upgrade Matrix --> RPC AIO.

This commit introduces JJB-Upgrade-Matrix which is a single job that
replaces all three previous jobs. The parallelisation of upgrades is
handled by a parallel block in a jenkins pipeline script rather than
seperate jobs. This new way of allocating nodes is not compatible with
jenkins jcloud single use slaves, as slaves are not reliably terminated
on completion. This would leave unclean slaves in the pool ready to
accept new jobs and that would cause future jobs to fail the node
cleanliness test.

Some major consequences of this approach:
  * The AIO and Upgrade Matrix Jobs used to call the same RPC-AIO job to
    do the majority of the work (but with different parameters). Now
    these jobs are both self contained, in order to prevent
    duplication, common code has been moved into bash_gating_lib.sh

  * As the jclouds approach to allocating nodes doesn't work well with
    pipeline script, the node creation/teardown is now handled by bash
    functions in gating_bash_lib. A positive of this is that its now
    possible to specify per job whether the node should be cleaned up.

Connects rcbops/u-suk-dev#353