Push new links and os-releases to rpc-repo
When a successful build is completed, the 'links' and 'os-releases'
folder should be pushed up to rpc-repo so that the wheels that were
compiled may be re-used in future builds.

The 'links' folder is used as a simple index for pip during the
repo-build process. The subfolders in 'os-releases' are per-tag
simple indexes for pip to use when installing packages.

These folders contain links to the artifacts held in the 'pools'
folder, but the links provide access to them for their respective
uses.