Ensure that no unwanted leftovers are in the artifact repo
The artifacts stored after a successful build may execute multiple
times and it is possible that it may leave some parts behind which
are unnecessary or not applicable.

For example:
The os-releases folder in a build may refer to a newer version of
a python package which happens to not be pinned by OpenStack's
upper constraints. In the artifact copy the older version of the
package will still be linked in the tag's os-releases folder, thus
referncing two possible versions fot that tag. This, in turn, will
create problems when executing a build using these artifacts
because the wrong python package version may be used.

This patch ensures that the os-releases and venv folders produced
in the build are copied to rpc-repo exactly, but also ensures that
the folders for other tags are left alone.