Exclude links folder index file from artifacting
For branches before Newton an index file is generated per
repo-build which only includes the indexed files for that
build. This therefore means that subsequent builds for
other branches will not benefit from the pre-compiled
wheels.

This patch ensures that the manually generated index file
is never copied over, allowing the web server's autoindex
to generate the index dynamically.