Create instance w/ correct name
Currently, when we spin up an instance, we give it a name such as
JRAmpm_7_13519.  However, when you log into the instance it has a
hostname of jrampm-7-13519.  The problem here is that when we go to
deploy maas, we look for an entity with label jrampm-7-13519 which does
not exist.

This commit simply changes how the instance name is formed so that it's
aligned with the hostname that the instance will have once it's booted.