Use credentials for apt jobs
In the apt job, we currently use job variables and credentials.
This could be source of errors, and every credential should be
stored at the same place. This effectively moves the apt- jobs
to the RPC_REPO_(<protocol>|<type of usage>)_<scope>_(FILE|TEXT)
pattern, with jenkins global credentials.

It also introduces the repo host key for known host checking in
the credentials.