Use artifacts folder variable
This commit moves to use a jenkins variable to define the artifact
folder. The environment variables are available everywhere (bash,
ansible) and therefore it's easier to use it as single source of
truth: if the folder location changes, we have one variable to
change in jenkins, instead of fixing the path in different repos
(like rpc-artifacts, jenkins-rpc).