Check resources survive an upgrade
This is a basic implementation of a check for resource status after an
upgrade.

A pair of instances are created before an upgrade,
then checked for existence and reachability after the upgrade.

Future improvement would be to measure the downtime of these instances
during the upgrade (using something like downtimer).

Connects rcbops/u-suk-dev#230
commit 10662d32d985eb87f63cfa1abb263f32a5365ca9