Provide a default for RPC_REPO
For the git artifacts job, the substitution for RPC_REPO
isn't working. Considering that this is a low impact
issue, it's simpler to just provide a default value
than to try and figure out what's going on in a system
that's about to be replaced.

Connects https://github.com/rcbops/u-suk-dev/issues/1111